move default colors to constants.js; refactor ResponsiveDonutWithInnerPercent.jsx
