+++ frontend/src/components/UI/ResponsiveDonutWithInnerPercent/CustomLayerComponent.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+
+const CustomLayerComponent = (myProps) => (layerProps) => {
+    const { centerX, centerY } = layerProps;
+
+    return (
+        <text
+            x={centerX}
+            y={centerY}
+            textAnchor="middle"
+            dominantBaseline="central"
+            style={{
+                fontSize: "20px",
+                fontWeight: "600",
+            }}
+        >
+            {myProps}
+        </text>
+    );
+};
+
+export default CustomLayerComponent;
