+++ frontend/src/components/UI/ResponsiveDonutWithInnerPercent/ResponsiveDonutWithInnerPercent.jsx
