+++ frontend/src/constants.js
@@ -1,5 +1,17 @@
 const constants = {
     notFilledInText: "--",
+    colors: [
+        "#336A90",
+        "#A1D0BE",
+        "#B50909",
+        "#E5A000",
+        "#6F3331",
+        "#C07B96",
+        "#264A64",
+        "#3E8D61",
+        "#D67625",
+        "#429195",
+    ],
 };
 
 export default constants;
